pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mitmproxy-adblock"
version = "0.2.0"
description = "Ad blocker for mitmproxy"
authors = ["Dominik Kozaczko <dominik@kozaczko.info>"]
license = "Apache License 2.0"
readme = "README.md"
repository = "https://github.com/dekoza/mitmproxy-adblock"
homepage = "https://github.com/dekoza/mitmproxy-adblock"
keywords = []
packages = [ { include = "mitm_adblock", from = "src" } ]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.4.0"
mitmproxy = "^7.0.2"
pyre2 = "^0.3.6"
adblockparser = "^0.7"

[tool.poetry.group.dev.dependencies]
# formatting, quality, tests
autoflake = "^1.4"
black = "^20.8b1"
isort = "^5.7.0"
mypy = "^0.812"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
pytest-randomly = "^3.5.0"
pytest-sugar = "^0.9.4"
pytest-xdist = "^2.2.0"

# tasks
duty = "^0.7.0"
git-changelog = "^0.4.2"
httpx = "^0.16.1"
jinja2-cli = "^0.7.0"
toml = "^0.10.2"

# flake8 plugins
darglint = "^1.5.8"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.1"
flake8-bugbear = "^20.11.1"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.3.1"
flake8-docstrings = "^1.5.0"
flake8-pytest-style = "^1.3.0"
flake8-string-format = "^0.3.0"
flake8-tidy-imports = "^4.2.1"
flake8-variables-names = "^0.0.4"
pep8-naming = "^0.11.1"
wps-light = "^0.15.2"

# docs
mkdocs = "^1.1.2"
mkdocs-coverage = "^0.2.1"
mkdocs-macros-plugin = "^0.5.0"
mkdocs-material = "^6.2.7"
mkdocstrings = "^0.15.0"

[tool.poetry.scripts]
mitm_adblock = "mitm_adblock.cli:main"

[tool.black]
line-length = 120
exclude = "tests/fixtures"

[tool.isort]
line_length = 120
not_skip = "__init__.py"
multi_line_output = 3
force_single_line = false
balanced_wrapping = true
default_section = "THIRDPARTY"
known_first_party = "mitm_adblock"
include_trailing_comma = true
